+++ test/testtransportation.jl
@@ -77,7 +77,7 @@
         ]
     end
 
-    @testset "Solution - 2" begin 
+    @testset "Solution - 2" begin
         t = TransportationProblem(
             Float64[10 20 25; 12 15 7],
             Float64[190, 40, 20],
@@ -87,7 +87,7 @@
 
         @test result.cost == 2720.0
         @test result.solution == [150.0 0.0 0.0; 40.0 40.0 20.0]
-    end 
+    end
 
     @testset "North-West Corner" begin
         @testset "Example 1" begin
@@ -134,9 +134,49 @@
             @test result.cost == 3510
         end
     end
-    
+
+
+
+
+    @testset "Least Cost Method" begin
+        @testset "Example 1" begin
+            t = TransportationProblem(
+                Float64[1 2 8 10; 7 3 4 5; 12 11 9 6],
+                Float64[50, 60, 30, 130],
+                Float64[90, 110, 70]
+            )
+            result = leastcost(t)
+            expected = Float64[
+                50 40 0 0;
+                0 20 30 60;
+                0 0 0 70
+            ]
+            @test result.solution == expected
+            @test result.cost == 1030
+        end
+
+        @testset "Example 2" begin
+            t = TransportationProblem(
+                Float64[10 2 20 11; 12 7 9 20; 4 14 16 18],
+                Float64[5, 15, 15, 15],
+                Float64[15, 25, 10]
+            )
+            
+            result = leastcost(t)
+            
+            expected = Float64[0 15 0 0; 0 0 15 10; 5 0 0 5]
+            
+            @test result.solution == expected
+            
+            @test result.cost == 475
+        end
+    end
+
+
+
+
     @testset "Big Example - Check if it is solved in reasonable times" verbose = true begin
-        
+
         t = TransportationProblem(
             rand(10:1000, 350, 450) * 1.0,
             rand(1:100, 450) * 1.0,
@@ -145,7 +185,7 @@
 
 
         result = solve(t)
-        
+
         @test result.cost > 0
 
     end